ANSI port

+++ include/minix/type.h
@@ -42,16 +42,16 @@ typedef int signed_clicks;	/* same length as phys_clicks, but signed */
 #define M4                 4
 #define M3_STRING         14
 
-typedef struct {int m1i1, m1i2, m1i3; char *m1p1, *m1p2, *m1p3;} mess_1;
-typedef struct {int m2i1, m2i2, m2i3; long m2l1, m2l2; char *m2p1;} mess_2;
-typedef struct {int m3i1, m3i2; char *m3p1; char m3ca1[M3_STRING];} mess_3;
+typedef struct {short m1i1, m1i2, m1i3; char *m1p1, *m1p2, *m1p3;} mess_1;
+typedef struct {short m2i1, m2i2, m2i3; long m2l1, m2l2; char *m2p1;} mess_2;
+typedef struct {short m3i1, m3i2; char *m3p1; char m3ca1[M3_STRING];} mess_3;
 typedef struct {long m4l1, m4l2, m4l3, m4l4;} mess_4;
 typedef struct {char m5c1, m5c2; int m5i1, m5i2; long m5l1, m5l2, m5l3;} mess_5;
-typedef struct {int m6i1, m6i2, m6i3; long m6l1; void (*m6f1)();} mess_6;
+typedef struct {short m6i1, m6i2, m6i3; long m6l1; void (*m6f1)();} mess_6;
 
 typedef struct {
-  int m_source;			/* who sent the message */
-  int m_type;			/* what kind of message is it */
+  short m_source;		/* who sent the message */
+  short m_type;			/* what kind of message is it */
   union {
 	mess_1 m_m1;
 	mess_2 m_m2;
